+++ subsys/net/ip/CMakeLists.txt
@@ -31,7 +31,6 @@ zephyr_library_sources(net_tc.c)
 zephyr_library_sources(icmp.c)
 zephyr_library_sources_ifdef(CONFIG_NET_IP           connection.c)
 zephyr_library_sources_ifdef(CONFIG_NET_6LO          6lo.c)
-zephyr_library_sources_ifdef(CONFIG_NET_DHCPV4       dhcpv4.c)
 zephyr_library_sources_ifdef(CONFIG_NET_DHCPV6       dhcpv6.c)
 zephyr_library_sources_ifdef(CONFIG_NET_IPV4_AUTO    ipv4_autoconf.c)
 zephyr_library_sources_ifdef(CONFIG_NET_IPV4         icmpv4.c ipv4.c)
