+++ subsys/net/lib/CMakeLists.txt
@@ -15,7 +15,10 @@ add_subdirectory_ifdef(CONFIG_NET_CAPTURE            capture)
 add_subdirectory_ifdef(CONFIG_NET_ZPERF              zperf)
 add_subdirectory_ifdef(CONFIG_NET_SHELL              shell)
 add_subdirectory_ifdef(CONFIG_NET_TRICKLE            trickle)
-add_subdirectory_ifdef(CONFIG_NET_DHCPV4_SERVER      dhcpv4)
+
+if (CONFIG_NET_DHCPV4 OR CONFIG_NET_DHCPV4_SERVER)
+  add_subdirectory(dhcpv4)
+endif()
 
 if (CONFIG_DNS_RESOLVER
     OR CONFIG_MDNS_RESPONDER
