+++ subsys/net/lib/dhcpv4/CMakeLists.txt
@@ -6,6 +6,5 @@ zephyr_library()
 
 zephyr_library_include_directories(${ZEPHYR_BASE}/subsys/net/ip)
 
-zephyr_library_sources(
-  dhcpv4_server.c
-  )
+zephyr_library_sources_ifdef(CONFIG_NET_DHCPV4 dhcpv4.c)
+zephyr_library_sources_ifdef(CONFIG_NET_DHCPV4_SERVER dhcpv4_server.c)
