+++ tests/net/dhcpv4/server/CMakeLists.txt
@@ -6,4 +6,5 @@ find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project(dhcpv4_server)
 
 target_include_directories(app PRIVATE ${ZEPHYR_BASE}/subsys/net/ip)
+target_include_directories(app PRIVATE ${ZEPHYR_BASE}/subsys/net/lib)
 target_sources(app PRIVATE src/main.c)
